# common/openapi.yaml
openapi: 3.0.0
info:
  title: Treasury Sandbox API
  description: API for the Confidential Treasury Strategy Sandbox - a privacy-preserving DAO treasury experimentation dApp
  version: 1.0.0

tags:
  - name: Config
  - name: Party
  - name: Epoch
  - name: Strategies
  - name: Performance
  - name: Governance
  - name: Demo
  - name: Auth
  - name: Feature Flags

paths:
  /feature-flags:
    get:
      tags: [Feature Flags]
      summary: Get feature flags
      operationId: getFeatureFlags
      responses:
        '200':
          description: A list of feature flags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlags'
        '500':
          $ref: '#/components/responses/InternalError'

  /login-links:
    get:
      tags: [Auth]
      summary: Get list of links that initiate login
      operationId: listLinks
      responses:
        '200':
          description: Links that initiate login flows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoginLink'
        '500':
          $ref: '#/components/responses/InternalError'

  /user:
    get:
      tags: [Auth]
      summary: Get Authenticated User
      operationId: getAuthenticatedUser
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /config:
    get:
      tags: [Config]
      summary: Get DAO configuration with all roles
      operationId: getDAOConfig
      responses:
        '200':
          description: DAO configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DAOConfig'
        '500':
          $ref: '#/components/responses/InternalError'

  /current-party:
    get:
      tags: [Party]
      summary: Get current party context
      operationId: getCurrentParty
      responses:
        '200':
          description: Current party info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyInfo'
        '500':
          $ref: '#/components/responses/InternalError'

  /party/switch:
    post:
      tags: [Party]
      summary: Switch active party context
      operationId: switchParty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchPartyRequest'
      responses:
        '200':
          description: Party switched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /epoch:
    get:
      tags: [Epoch]
      summary: Get current epoch state
      operationId: getEpochState
      responses:
        '200':
          description: Current epoch state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpochState'
        '500':
          $ref: '#/components/responses/InternalError'

  /epoch/advance:
    post:
      tags: [Epoch]
      summary: Advance to the next epoch
      operationId: advanceEpoch
      responses:
        '200':
          description: Epoch advanced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpochState'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /strategies:
    get:
      tags: [Strategies]
      summary: List strategies (filtered by party visibility)
      operationId: listStrategies
      responses:
        '200':
          description: List of strategies visible to current party
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Strategy'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Strategies]
      summary: Create a new strategy
      operationId: createStrategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrategyRequest'
      responses:
        '201':
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /strategies/{strategyId}/allocations:
    put:
      tags: [Strategies]
      summary: Update strategy allocations
      operationId: updateAllocations
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAllocationsRequest'
      responses:
        '200':
          description: Allocations updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /performance:
    get:
      tags: [Performance]
      summary: Get all performance reports
      operationId: listPerformanceReports
      responses:
        '200':
          description: List of performance reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PerformanceReport'
        '500':
          $ref: '#/components/responses/InternalError'

  /performance/publish:
    post:
      tags: [Performance]
      summary: Publish performance reports for current epoch
      operationId: publishPerformance
      responses:
        '201':
          description: Performance reports published
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PerformanceReport'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /votes/{epoch}:
    get:
      tags: [Governance]
      summary: Get votes for a specific epoch
      operationId: getVotes
      parameters:
        - name: epoch
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Votes for the epoch
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vote'
        '500':
          $ref: '#/components/responses/InternalError'

  /votes:
    post:
      tags: [Governance]
      summary: Cast an elimination vote
      operationId: castVote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CastVoteRequest'
      responses:
        '201':
          description: Vote cast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /elimination/execute:
    post:
      tags: [Governance]
      summary: Execute elimination based on votes
      operationId: executeElimination
      responses:
        '200':
          description: Elimination executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EliminationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /eliminations:
    get:
      tags: [Governance]
      summary: Get elimination history
      operationId: listEliminations
      responses:
        '200':
          description: Elimination history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EliminationResult'
        '500':
          $ref: '#/components/responses/InternalError'

  /demo/seed:
    post:
      tags: [Demo]
      summary: Seed demo data
      operationId: seedDemo
      responses:
        '201':
          description: Demo data seeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeedResult'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  parameters:
    ContractId:
      name: contractId
      in: path
      required: true
      description: Ledger contract ID
      schema:
        type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Not authenticated
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: An internal server error occurred
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    FeatureFlags:
      type: object
      properties:
        authMode:
          type: string
          enum:
            - oauth2
            - shared-secret

    LoginLink:
      type: object
      required:
        - name
        - url
      properties:
        name:
          type: string
        url:
          type: string

    AuthenticatedUser:
      type: object
      required:
        - name
        - party
        - roles
        - isAdmin
        - walletUrl
      properties:
        name:
          type: string
        party:
          type: string
        roles:
          type: array
          items:
            type: string
        isAdmin:
          type: boolean
        walletUrl:
          type: string

    DAOConfig:
      type: object
      required:
        - operator
        - strategyManager
        - voters
        - auditor
        - publicObserver
      properties:
        operator:
          type: string
        strategyManager:
          type: string
        voters:
          type: array
          items:
            type: string
        auditor:
          type: string
        publicObserver:
          type: string

    PartyInfo:
      type: object
      required:
        - partyId
        - role
        - displayName
      properties:
        partyId:
          type: string
        role:
          type: string
          enum:
            - operator
            - strategyManager
            - voter
            - auditor
            - publicObserver
        displayName:
          type: string

    SwitchPartyRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum:
            - operator
            - strategyManager
            - voter1
            - voter2
            - voter3
            - auditor
            - publicObserver

    EpochState:
      type: object
      required:
        - currentEpoch
        - totalEpochs
        - phase
      properties:
        currentEpoch:
          type: integer
        totalEpochs:
          type: integer
        phase:
          type: string
          enum:
            - Reporting
            - Voting
            - Completed
        contractId:
          type: string

    Strategy:
      type: object
      required:
        - strategyId
        - name
        - riskCategory
        - status
        - isAllocationsVisible
      properties:
        strategyId:
          type: string
        name:
          type: string
        riskCategory:
          type: string
          enum:
            - Conservative
            - Moderate
            - Aggressive
        allocations:
          $ref: '#/components/schemas/Allocations'
        status:
          type: string
          enum:
            - Active
            - Eliminated
        isAllocationsVisible:
          type: boolean
          description: Whether the current party can see allocations
        contractId:
          type: string

    Allocations:
      type: object
      properties:
        ethWeight:
          type: number
          format: double
        btcWeight:
          type: number
          format: double
        usdcWeight:
          type: number
          format: double

    CreateStrategyRequest:
      type: object
      required:
        - name
        - riskCategory
        - allocations
      properties:
        name:
          type: string
        riskCategory:
          type: string
          enum:
            - Conservative
            - Moderate
            - Aggressive
        allocations:
          $ref: '#/components/schemas/Allocations'

    UpdateAllocationsRequest:
      type: object
      required:
        - allocations
      properties:
        allocations:
          $ref: '#/components/schemas/Allocations'

    PerformanceReport:
      type: object
      required:
        - strategyId
        - strategyName
        - riskCategory
        - epoch
        - epochReturn
        - cumulativeReturn
        - maxDrawdown
      properties:
        strategyId:
          type: string
        strategyName:
          type: string
        riskCategory:
          type: string
        epoch:
          type: integer
        epochReturn:
          type: number
          format: double
        cumulativeReturn:
          type: number
          format: double
        maxDrawdown:
          type: number
          format: double

    Vote:
      type: object
      required:
        - voter
        - epoch
        - targetStrategyId
      properties:
        voter:
          type: string
        epoch:
          type: integer
        targetStrategyId:
          type: string
        contractId:
          type: string

    CastVoteRequest:
      type: object
      required:
        - targetStrategyId
      properties:
        targetStrategyId:
          type: string

    EliminationResult:
      type: object
      required:
        - epoch
        - eliminatedStrategyId
        - eliminatedStrategyName
        - voteTally
      properties:
        epoch:
          type: integer
        eliminatedStrategyId:
          type: string
        eliminatedStrategyName:
          type: string
        voteTally:
          type: object
          additionalProperties:
            type: integer
        contractId:
          type: string

    SeedResult:
      type: object
      properties:
        message:
          type: string
        strategiesCreated:
          type: integer
        epochInitialized:
          type: boolean

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          nullable: true